Add a kitchen sink
Closes #224
