+++ iopipe/__init__.py
@@ -1,1 +1,2 @@
+from .agent import IOpipeCore  # noqa
 from .iopipe import IOpipe  # noqa
